+++ src/iqm/qiskit_iqm/fake_backends/__init__.py
@@ -15,6 +15,7 @@
 """
 
 from .fake_adonis import IQMFakeAdonis
+from .fake_aphrodite import IQMFakeAphrodite
 from .fake_apollo import IQMFakeApollo
 from .fake_deneb import IQMFakeDeneb
 from .iqm_fake_backend import IQMErrorProfile, IQMFakeBackend